requirements.txt
# requirements.txt

# Web framework
fastapi==0.110.3
uvicorn[standard]==0.30.0
email-validator==2.1.1

# Configuration
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Migrations
alembic==1.13.1

# Security / Auth
python-jose[cryptography]==3.3.0  # for JWT encoding/decoding
passlib[bcrypt]==1.7.4  # for secure password hashing

# CORS
python-multipart==0.0.9  # for file uploads if needed

# Testing
pytest==8.2.1
pytest-cov==5.0.0
httpx==0.27.0  # for async API test client if needed

# Linting / formatting (optional but highly recommended)
black==24.4.2
isort==5.13.2

# Resume parsing (you mentioned this in your plan)
PyPDF2==3.0.1
python-docx==1.1.2
spacy==3.7.4  # optional for advanced parsing

# Vector database
pgvector==0.2.4

# Added from the code block
openai==1.12.0
